# public/openapi.yml
openapi: 3.0.3
info:
  version: "1"
  title: "Food delivery API"
  description: "Demo Food delivery API"
paths:
...
  /restaurant:
    post:
      summary: Create
      operationId: createRestaurant
      tags:
        - Restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantRequestBody"
      responses:
        200:
          description: Restaurant Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantSchema"
    get:
      summary: Get Restaurants
      operationId: getRestaurants
      tags:
        - Restaurant
      parameters:
        - in: query
          name: filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommonFilter"
      responses:
        200:
          description: Restaurants Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantsResponse"

    delete:
      summary: Delete all Restaurants
      tags:
        - Restaurant
      operationId: deleteAllRestaurants
      responses:
        204:
          description: Success
  /restaurant/{id}:
    patch:
      summary: Update
      operationId: updateRestaurantById
      tags:
        - Restaurant
      parameters:
        - name: id
          in: path
          description: Restaurant ID
          required: true
          schema:
            type: number
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantRequestBody"
      responses:
        200:
          description: Restaurant Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantSchema"
    get:
      summary: Get Restaurant by ID
      operationId: getRestaurantById
      tags:
        - Restaurant
      parameters:
        - name: id
          in: path
          description: Restaurant ID
          required: true
          schema:
            type: number
          example: 1
      responses:
        200:
          description: Restaurant Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantSchema"
    delete:
      summary: Delete Restaurant by ID
      tags:
        - Restaurant
      operationId: deleteOneRestaurant
      parameters:
        - name: id
          in: path
          description: Restaurant ID
          required: true
          schema:
            type: number
          example: 1
      responses:
        204:
          description: Success
  # END Restaurant

  # Courier
...
components:
  schemas:
    RestaurantSchema:
      description: Restaurant schema
      type: object
      required:
        - id
        - name
        - picture
        - created_at
        - updated_at
      properties:
        id:
          type: number
        name:
          type: string
        picture:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      example:
        id: 1
        name: Some Restaurant
        picture: some_pic_restaurant
        created_at: '2020-02-10 17:32:28'
        updated_at: '2020-02-10 17:32:28'
    RestaurantRequestBody:
      description: Create Restaurant
      type: object
      required:
        - name
        - picture
      properties:
        name:
          type: string
        picture:
          type: string
      example:
        name: Some Restaurant
        picture: some_pic_restaurant
    RestaurantsResponse:
      description: List of Restaurants
      type: array
      items:
        $ref: '#/components/schemas/RestaurantSchema'
...
    CommonFilter:
      type: object
      properties:
        offset:
          type: integer
          minimum: 0
          example: 0
        limit:
          type: integer
          minimum: 1
          example: 100
        order_by:
          type: object
          properties:
            sort_direction:
              type: string
              enum: [ASC, DESC]
              example: ASC
            field:
              type: string
              example: created_at